Move area endpoints to an independent file.
