+++ api/processor_api.py
@@ -8,17 +8,17 @@ from fastapi.staticfiles import StaticFiles
 from fastapi.exceptions import RequestValidationError
 from fastapi.responses import JSONResponse
 from pydantic import BaseModel
-from starlette.exceptions import HTTPException
 from share.commands import Commands
 from typing import Optional
 
 from .cameras import cameras_api, map_camera, map_to_camera_file_format
-from .models.config_keys import AreaConfigDTO, ConfigDTO
+from .areas import areas_api, map_area, map_to_area_file_format
+from .models.config_keys import ConfigDTO
 from .queue_manager import QueueManager
 from .reports import reports_api
 from .settings import Settings
 from .utils import (
-    extract_config, handle_config_response, reestructure_areas, update_and_restart_config
+    extract_config, handle_config_response, update_and_restart_config
 )
 
 logger = logging.getLogger(__name__)
@@ -62,6 +62,7 @@ class ProcessorAPI:
 
         app.mount("/reports", reports_api)
         app.mount("/cameras", cameras_api)
+        app.mount("/areas", areas_api)
 
         @app.exception_handler(RequestValidationError)
         async def validation_exception_handler(request: Request, exc: RequestValidationError):
@@ -81,35 +82,13 @@ class ProcessorAPI:
 
         app.mount("/static", StaticFiles(directory="/repo/data/processor/static"), name="static")
 
-        def map_area(area_name, config):
-            area = config.get(area_name)
-
-            return {
-                "id": area.get("Id"),
-                "name": area.get("Name"),
-                "cameras": area.get("Cameras"),
-                "notifyEveryMinutes": area.get("NotifyEveryMinutes"),
-                "emails": area.get("Emails"),
-                "occupancyThreshold": area.get("OccupancyThreshold"),
-                "violationThreshold": area.get("ViolationThreshold"),
-                "dailyReport": area.get('DailyReport'),
-                'dailyReportTime': area.get('DailyReportTime')
-            }
-
-        def map_to_area_file_format(area: AreaConfigDTO):
-            return dict(
-                {
-                    'Id': area.id,
-                    'Name': area.name,
-                    'Cameras': area.cameras,
-                    'NotifyEveryMinutes': str(area.notifyEveryMinutes),
-                    'Emails': area.emails,
-                    'OccupancyThreshold': str(area.occupancyThreshold),
-                    'ViolationThreshold': str(area.violationThreshold),
-                    "DailyReport": str(area.dailyReport),
-                    'DailyReportTime': area.dailyReportTime
-                }
-            )
+        def map_to_config_file_format(config_dto):
+            config_dict = dict()
+            for count, camera in enumerate(config_dto.cameras):
+                config_dict["Source_" + str(count)] = map_to_camera_file_format(camera)
+            for count, area in enumerate(config_dto.areas):
+                config_dict["Area_" + str(count)] = map_to_area_file_format(area)
+            return config_dict
 
         def map_config(config, options):
             cameras_name = [x for x in config.keys() if x.startswith("Source")]
@@ -120,14 +99,6 @@ class ProcessorAPI:
                 "cameras": [map_camera(x, config, options) for x in cameras_name],
                 "areas": [map_area(x, config) for x in areas_name]
             }
-
-        def map_to_config_file_format(config_dto):
-            config_dict = dict()
-            for count, camera in enumerate(config_dto.cameras):
-                config_dict["Source_" + str(count)] = map_to_camera_file_format(camera)
-            for count, area in enumerate(config_dto.areas):
-                config_dict["Area_" + str(count)] = map_to_area_file_format(area)
-            return config_dict
 
         def write_user_token(token):
             logger.info("Writing user access token")
@@ -160,10 +131,6 @@ class ProcessorAPI:
             success = update_and_restart_config(config_dict)
             return handle_config_response(config_dict, success)
 
-        def get_areas():
-            config = extract_config(config_type='areas')
-            return [map_area(x, config) for x in config.keys()]
-
         @app.get("/process-video-cfg")
         async def process_video_cfg():
             logger.info("process-video-cfg requests on api")
@@ -188,80 +155,6 @@ class ProcessorAPI:
         @app.put("/config")
         async def update_config(config: ConfigDTO):
             config_dict = map_to_config_file_format(config)
-
-            success = update_and_restart_config(config_dict)
-            return handle_config_response(config_dict, success)
-
-        @app.get("/areas")
-        async def list_areas():
-            return {
-                "areas": get_areas()
-            }
-
-        @app.get("/areas/{area_id}")
-        async def get_area(area_id):
-            area = next((area for area in get_areas() if area['id'] == area_id), None)
-            if not area:
-                raise HTTPException(status_code=404, detail=f'The area: {area_id} does not exist')
-            return area
-
-        @app.post('/areas')
-        async def create_area(new_area: AreaConfigDTO):
-            config_dict = extract_config()
-            areas_name = [x for x in config_dict.keys() if x.startswith("Area")]
-            areas = [map_area(x, config_dict) for x in areas_name]
-            if new_area.id in [area['id'] for area in areas]:
-                raise HTTPException(status_code=400, detail="Area already exists")
-
-            cameras = [x for x in config_dict.keys() if x.startswith("Source")]
-            cameras = [map_camera(x, config_dict, []) for x in cameras]
-            camera_ids = [camera['id'] for camera in cameras]
-            if not all(x in camera_ids for x in new_area.cameras.split(',')):
-                non_existent_cameras = set(new_area.cameras.split(',')) - set(camera_ids)
-                raise HTTPException(status_code=404, detail=f'The cameras: {non_existent_cameras} do not exist')
-
-            config_dict[f'Area_{len(areas)}'] = map_to_area_file_format(new_area)
-
-            success = update_and_restart_config(config_dict)
-            return handle_config_response(config_dict, success)
-
-        @app.put('/areas/{area_id}')
-        async def edit_area(area_id, edited_area: AreaConfigDTO):
-            edited_area.id = area_id
-            config_dict = extract_config()
-            area_names = [x for x in config_dict.keys() if x.startswith("Area")]
-            areas = [map_area(x, config_dict) for x in area_names]
-            areas_ids = [area['id'] for area in areas]
-            try:
-                index = areas_ids.index(area_id)
-            except ValueError:
-                raise HTTPException(status_code=404, detail=f'The area: {area_id} does not exist')
-
-            cameras = [x for x in config_dict.keys() if x.startswith("Source")]
-            cameras = [map_camera(x, config_dict, []) for x in cameras]
-            camera_ids = [camera['id'] for camera in cameras]
-            if not all(x in camera_ids for x in edited_area.cameras.split(',')):
-                non_existent_cameras = set(edited_area.cameras.split(',')) - set(camera_ids)
-                raise HTTPException(status_code=404, detail=f'The cameras: {non_existent_cameras} do not exist')
-
-            config_dict[f"Area_{index}"] = map_to_area_file_format(edited_area)
-
-            success = update_and_restart_config(config_dict)
-            return handle_config_response(config_dict, success)
-
-        @app.delete('/areas/{area_id}')
-        async def delete_area(area_id):
-            config_dict = extract_config()
-            areas_name = [x for x in config_dict.keys() if x.startswith("Area")]
-            areas = [map_area(x, config_dict) for x in areas_name]
-            areas_ids = [area['id'] for area in areas]
-            try:
-                index = areas_ids.index(area_id)
-            except ValueError:
-                raise HTTPException(status_code=404, detail=f'The area: {area_id} does not exist')
-
-            config_dict.pop(f'Area_{index}')
-            config_dict = reestructure_areas((config_dict))
 
             success = update_and_restart_config(config_dict)
             return handle_config_response(config_dict, success)
